Update platform.c
Added new states for Knockback and Blank
Added some additional state initialization elements to help keep the change-state event clean (esp ground initialization). Cut ground_reset() entirely

added a check for coyote time count down so that it doesn't have to be reset each time. Though I'm not sure this is the cleanest OR the most efficient route.

Cut an extra clamp in the Y_vel check that wasn't doing anything because vel_y was being set to 0.

Added a new jump_type to keep track of floating